// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ReadStatus = {
  TO_READ: 'TO_READ',
  READING: 'READING',
  READ: 'READ',
  READ_NEXT: 'READ_NEXT',
  DNF: 'DNF'
} as const

export type ReadStatus = (typeof ReadStatus)[keyof typeof ReadStatus]